src/main/java/com/tw/hotelmenu/Service/UserService.java
package com.tw.hotelmenu.Service;

import com.tw.hotelmenu.Repository.UserRepository;
import com.tw.hotelmenu.Service.dto.UserDetailsDTO;
import lombok.AllArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@AllArgsConstructor
public class UserService implements UserDetailsService {

    UserRepository userRepository;
    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        return userRepository.findByEmail(username)
                .map(UserDetailsDTO::create)
                .orElseThrow(()->new UsernameNotFoundException(String.format("No User exists with Username : %s", username)));
    }
}
